requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
fhirclient==4.2.0
Werkzeug==3.0.1
sqlalchemy==2.0.23
gunicorn==21.2.0
psycopg2-binary==2.9.9
